+++ lib/src/features/entitlements/domain/repositories/entitlement_snapshot_source.dart
@@ -1,0 +1,5 @@
+import '../entities/entitlement_snapshot.dart';
+
+abstract interface class EntitlementSnapshotSource {
+  Future<EntitlementSnapshot> loadSnapshot();
+}